# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HalkYemek API
  version: dev
  description: Beylikdüzü odaklı QR + Wallet yemek sistemi
paths:
...
  /api/v1/ops/dashboard/:
    get:
      operationId: ops_payout_dashboard
      tags:
      - ops-payouts
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/ops/payouts/:
    get:
      operationId: ops_payouts_list
      tags:
      - ops-payouts
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/ops/payouts/{payout_id}/:
    get:
      operationId: ops_payouts_detail
      parameters:
      - in: path
        name: payout_id
        schema:
          type: integer
        required: true
      tags:
      - ops-payouts
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/ops/payouts/{payout_id}/confirm/:
    post:
      operationId: ops_payout_confirm
      parameters:
      - in: path
        name: payout_id
        schema:
          type: integer
        required: true
      tags:
      - ops-payouts
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/ops/payouts/dispatch-due/:
    post:
      operationId: ops_payout_dispatch_due
      tags:
      - ops-payouts
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/ops/reconcile/business/{business_id}/:
    get:
      operationId: ops_reconcile_business
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      tags:
      - ops-payouts
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/wallet/:
    get:
      operationId: wallet_detail
      tags:
      - wallet
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
              examples:
                WalletDetail:
                  value:
                    user_id: 12
                    balance: 900
                    pending_balance: 0
                    is_active: true
                    ledger_in_sync: true
                    pending_ledger_in_sync: true
                    created_at: '2026-04-02T11:00:00+03:00'
                    updated_at: '2026-04-02T12:00:00+03:00'
                  summary: Wallet detail
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
  /api/v1/wallet/pending-transactions/:
    get:
      operationId: wallet_pending_transaction_list
      parameters:
      - name: ordering
        required: false
        in: query
        description: Sonuçların sıralanmasında kullanılacak alan.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Sayfalanmış sonuç kümesinde bir sayfa numarası.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Her sayfada döndürülecek sonuç sayısı.
        schema:
          type: integer
      - in: query
        name: payment_intent_id
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: Bir arama terimi.
        schema:
          type: string
      - in: query
        name: type
        schema:
          type: string
          minLength: 1
      tags:
      - wallet
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPendingWalletTransactionList'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
  /api/v1/wallet/transactions/:
    get:
      operationId: wallet_transaction_list
      parameters:
      - in: query
        name: order_id
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Sonuçların sıralanmasında kullanılacak alan.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Sayfalanmış sonuç kümesinde bir sayfa numarası.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Her sayfada döndürülecek sonuç sayısı.
        schema:
          type: integer
      - in: query
        name: payment_intent_id
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: Bir arama terimi.
        schema:
          type: string
      - in: query
        name: type
        schema:
          type: string
          minLength: 1
      tags:
      - wallet
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWalletTransactionList'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
components:
  schemas:
...
    ApiErrorBody:
      type: object
      properties:
        code:
          type: string
        message: {}
        request_id:
          type: string
        details: {}
        reason:
          type: string
      required:
      - code
      - message
      - request_id
    ApiErrorEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          default: false
        error:
          $ref: '#/components/schemas/ApiErrorBody'
      required:
      - error
...
    PaginatedPendingWalletTransactionList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/PendingWalletTransaction'
    PaginatedWalletTransactionList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/WalletTransaction'
...
    PendingWalletTransaction:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        transaction_type:
          allOf:
          - $ref: '#/components/schemas/PendingWalletTransactionTransactionTypeEnum'
          readOnly: true
        amount:
          type: integer
          readOnly: true
          description: Signed kuruş. TOPUP_PENDING:+, SETTLEMENT_OUT:-
        before_pending:
          type: integer
          readOnly: true
        after_pending:
          type: integer
          readOnly: true
        provider_event_id:
          type: string
          readOnly: true
          nullable: true
        payment_intent_id:
          type: integer
          readOnly: true
          nullable: true
        description:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - after_pending
      - amount
      - before_pending
      - created_at
      - description
      - id
      - payment_intent_id
      - provider_event_id
      - transaction_type
    PendingWalletTransactionTransactionTypeEnum:
      enum:
      - TOPUP_PENDING
      - SETTLEMENT_OUT
      - REVERSAL_OUT
      type: string
      description: |-
        * `TOPUP_PENDING` - Topup Pending
        * `SETTLEMENT_OUT` - Settlement Out
        * `REVERSAL_OUT` - Reversal Out
...
    Wallet:
      type: object
      properties:
        user_id:
          type: integer
          readOnly: true
        balance:
          type: integer
          readOnly: true
          description: Kuruş cinsinden kullanılabilir bakiye
        pending_balance:
          type: integer
          readOnly: true
          description: Kuruş cinsinden settlement bekleyen bakiye
        is_active:
          type: boolean
          readOnly: true
        restriction_reason:
          type: string
          readOnly: true
        restricted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        ledger_in_sync:
          type: boolean
          readOnly: true
        pending_ledger_in_sync:
          type: boolean
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - balance
      - created_at
      - is_active
      - ledger_in_sync
      - pending_balance
      - pending_ledger_in_sync
      - restricted_at
      - restriction_reason
      - updated_at
      - user_id
    WalletTransaction:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        transaction_type:
          allOf:
          - $ref: '#/components/schemas/WalletTransactionTransactionTypeEnum'
          readOnly: true
        amount:
          type: integer
          readOnly: true
          description: Kuruş cinsinden signed işlem tutarı
        before_balance:
          type: integer
          readOnly: true
          description: İşlem öncesi bakiye
        after_balance:
          type: integer
          readOnly: true
          description: İşlem sonrası bakiye
        order_id:
          type: integer
          readOnly: true
          nullable: true
        provider_event_id:
          type: string
          readOnly: true
          nullable: true
        payment_intent_id:
          type: integer
          readOnly: true
          nullable: true
        description:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - after_balance
      - amount
      - before_balance
      - created_at
      - description
      - id
      - order_id
      - payment_intent_id
      - provider_event_id
      - transaction_type
    WalletTransactionTransactionTypeEnum:
      enum:
      - TOP_UP
      - PURCHASE
      - REFUND
      - ADJUSTMENT
      - REVERSAL
      - CHARGEBACK
      type: string
      description: |-
        * `TOP_UP` - Top Up
        * `PURCHASE` - Purchase
        * `REFUND` - Refund
        * `ADJUSTMENT` - Adjustment
        * `REVERSAL` - Reversal
        * `CHARGEBACK` - Chargeback
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...